Move code from Ribbon1 to Action Framework
