+++ PowerPointLabs/PowerPointLabs/ActionFramework/Label/FrostedGlassLabelHandler.cs
@@ -1,0 +1,14 @@
+﻿using PowerPointLabs.ActionFramework.Common.Attribute;
+using PowerPointLabs.ActionFramework.Common.Interface;
+
+namespace PowerPointLabs.ActionFramework.Label
+{
+    [ExportLabelRibbonId("frostedGlassEffect")]
+    class FrostedGlassLabelHandler : LabelHandler
+    {
+        protected override string GetLabel(string ribbonId)
+        {
+            return TextCollection.EffectsLabFrostedGlassButtonLabel;
+        }
+    }
+}
