+++ PowerPointLabs/PowerPointLabs/ActionFramework/Supertip/FrostedGlassSupertipHandler.cs
@@ -1,0 +1,14 @@
+﻿using PowerPointLabs.ActionFramework.Common.Attribute;
+using PowerPointLabs.ActionFramework.Common.Interface;
+
+namespace PowerPointLabs.ActionFramework.Supertip
+{
+    [ExportSupertipRibbonId("frostedGlassEffect")]
+    class FrostedGlassSupertipHandler : SupertipHandler
+    {
+        protected override string GetSupertip(string ribbonId)
+        {
+            return TextCollection.EffectsLabFrostedGlassSupertip;
+        }
+    }
+}
