+++ PowerPointLabs/PowerPointLabs/Ribbon1.cs
@@ -2133,22 +2133,6 @@ namespace PowerPointLabs
             TransparentEffect(shapeRange);
         }
 
-        public void FrostedGlassEffectClick(Office.IRibbonControl control)
-        {
-            Globals.ThisAddIn.Application.StartNewUndoEntry();
-            
-            var selection = PowerPointCurrentPresentationInfo.CurrentSelection;
-
-            if (!(selection.Type == PowerPoint.PpSelectionType.ppSelectionShapes
-                || selection.Type == PowerPoint.PpSelectionType.ppSelectionText))
-            {
-                MessageBox.Show(TextCollection.EffectsLabErrorFrostedGlassNoSelection, "Error");
-                return;
-            }
-
-            FrostedGlassEffect(selection.ShapeRange);
-        }
-
         private void MagnifyGlassEffect(PowerPoint.Shape shape, float ratio)
         {
             var delta = 0.5f * (ratio - 1);
@@ -2318,269 +2302,6 @@ namespace PowerPointLabs
         {
             shape.Fill.Transparency = 0.5f;
             shape.Line.Transparency = 0.5f;
-        }
-
-        private void FrostedGlassEffect(PowerPoint.ShapeRange shapeRange)
-        {
-            shapeRange.Cut();
-
-            var imageFile = Path.Combine(Path.GetTempPath(), "blur.png");
-            Utils.Graphics.ExportSlide(PowerPointCurrentPresentationInfo.CurrentSlide, imageFile);
-            var blurSlideShape = GenerateBlurShape(imageFile);
-            FitToSlide.AutoFit(blurSlideShape, PowerPointPresentation.Current.SlideWidth,
-                PowerPointPresentation.Current.SlideHeight);
-
-            shapeRange = PowerPointCurrentPresentationInfo.CurrentSlide.Shapes.Paste();
-
-            try
-            {
-                var ungroupedShapeRange = UngroupAllForFrostedGlassShapeRange(shapeRange);
-                var textBoxes = new List<PowerPoint.Shape>();
-                var blurShapeRange = GenerateFrostedGlassShapeRange(ungroupedShapeRange, ref textBoxes);
-                
-                blurSlideShape.ZOrder(Office.MsoZOrderCmd.msoBringToFront);
-                var blurShape = CropToShape.Crop(blurShapeRange, isInPlace: true, handleError: false);
-                blurSlideShape.Delete();
-
-                var overlayShape = GenerateOverlayShape(blurShape);
-                if (overlayShape.Type == Office.MsoShapeType.msoGroup)
-                {
-                    var overlayShapeRange = overlayShape.Ungroup();
-                    overlayShapeRange.MergeShapes(Office.MsoMergeCmd.msoMergeUnion);
-                }
-
-                foreach (var textBox in textBoxes)
-                {
-                    textBox.ZOrder(Office.MsoZOrderCmd.msoBringToFront);
-
-                    // prevent offset when cut and paste a shape that have another shape in the same location
-                    textBox.IncrementLeft(-12);
-                    textBox.IncrementTop(-12);
-                }
-            }
-            catch (Exception e)
-            {
-                blurSlideShape.Delete();
-
-                var messageBoxTitle = "Error";
-                var errorMessage = e.Message;
-                if (System.Text.RegularExpressions.Regex.IsMatch(errorMessage, @"^\d$"))
-                {
-                    errorMessage = CropToShape.GetErrorMessageForErrorCode(errorMessage);
-                }
-                switch (errorMessage)
-                {
-                    case TextCollection.CropToShapeText.ErrorMessageForSelectionNonShape:
-                        errorMessage = errorMessage.Insert(errorMessage.IndexOf("objects"), "and text box ");
-                        goto case TextCollection.EffectsLabErrorFrostedGlassEmptyTextBox;
-                    case TextCollection.EffectsLabErrorFrostedGlassEmptyTextBox:
-                        errorMessage = errorMessage.Replace("Crop To Shape", "Frosted Glass");
-                        MessageBox.Show(errorMessage, messageBoxTitle);
-                        break;
-                    default:
-                        ErrorDialogWrapper.ShowDialog(messageBoxTitle, e.Message, e);
-                        break;
-                }
-            }
-        }
-
-        private PowerPoint.ShapeRange UngroupAllForFrostedGlassShapeRange(PowerPoint.ShapeRange shapeRange)
-        {
-            var ungroupedShapeNames = new List<string>();
-            var queue = new Queue<PowerPoint.Shape>();
-
-            foreach (PowerPoint.Shape shape in shapeRange)
-            {
-                queue.Enqueue(shape);
-            }
-
-            while (queue.Count != 0)
-            {
-                var shape = queue.Dequeue();
-
-                switch (shape.Type)
-                {
-                    case Office.MsoShapeType.msoGroup:
-                        var subRange = shape.Ungroup();
-                        foreach (PowerPoint.Shape item in subRange)
-                        {
-                            queue.Enqueue(item);
-                        }
-                        break;
-                    case Office.MsoShapeType.msoPlaceholder:
-                    case Office.MsoShapeType.msoTextBox:
-                        if (String.IsNullOrWhiteSpace(shape.TextFrame2.TextRange.Text))
-                        {
-                            throw new Exception(TextCollection.EffectsLabErrorFrostedGlassEmptyTextBox);
-                        }
-                        goto case Office.MsoShapeType.msoFreeform;
-                    case Office.MsoShapeType.msoAutoShape:
-                    case Office.MsoShapeType.msoFreeform:
-                        ungroupedShapeNames.Add(shape.Name);
-                        break;
-                    default:
-                        throw new Exception(TextCollection.CropToShapeText.ErrorMessageForSelectionNonShape);
-                }
-            }
-
-            var ungroupedShapeRange = PowerPointCurrentPresentationInfo.CurrentSlide.Shapes.Range(ungroupedShapeNames.ToArray());
-
-            return ungroupedShapeRange;
-        }
-
-        private PowerPoint.ShapeRange GenerateFrostedGlassShapeRange(PowerPoint.ShapeRange shapeRange, ref List<PowerPoint.Shape> textBoxes)
-        {
-            var blurShapeNames = new List<string>();
-            var queue = new Queue<PowerPoint.Shape>();
-
-            foreach (PowerPoint.Shape shape in shapeRange)
-            {
-                if (shape.Type == Office.MsoShapeType.msoPlaceholder
-                    || shape.Type == Office.MsoShapeType.msoTextBox)
-                {
-                    if (!String.IsNullOrWhiteSpace(shape.TextFrame2.TextRange.Text))
-                    {
-                        var textBoundaryShape = GenerateShapeFromTextBoundary(shape);
-
-                        // prevent offset when cut and paste a shape that have another shape in the same location
-                        shape.IncrementLeft(12);
-                        shape.IncrementTop(12);
-
-                        textBoxes.Add(shape);
-                        blurShapeNames.Add(textBoundaryShape.Name);
-                    }
-                    else
-                    {
-                        throw new Exception(TextCollection.EffectsLabErrorFrostedGlassEmptyTextBox);
-                    }
-                }
-                else // if (shape.Type == Office.MsoShapeType.msoAutoShape || shape.Type == Office.MsoShapeType.msoFreeform)
-                {
-                    if (!String.IsNullOrWhiteSpace(shape.TextFrame2.TextRange.Text))
-                    {
-                        var textBox = DuplicateShapeInPlace(shape);
-                        textBox.Fill.Visible = Office.MsoTriState.msoFalse;
-                        textBox.Line.Visible = Office.MsoTriState.msoFalse;
-
-                        // prevent offset when cut and paste a shape that have another shape in the same location
-                        textBox.IncrementLeft(12);
-                        textBox.IncrementTop(12);
-
-                        textBoxes.Add(textBox);
-                        shape.TextFrame2.DeleteText();
-                    }
-
-                    blurShapeNames.Add(shape.Name);
-                }
-            }
-
-            var frostedGlassShapeRange = PowerPointCurrentPresentationInfo.CurrentSlide.Shapes.Range(blurShapeNames.ToArray());
-
-            return frostedGlassShapeRange;
-        }
-
-        private PowerPoint.Shape GenerateBlurShape(string imageFile)
-        {
-            using (var imageFactory = new ImageProcessor.ImageFactory())
-            {
-                var image = imageFactory
-                    .Load(imageFile)
-                    .Image;
-
-                var ratio = (float)image.Width / image.Height;
-                var targetHeight = Math.Round(1100f - (1100f - 11f) / 100f * 95);
-                var targetWidth = Math.Round(targetHeight * ratio);
-
-                image = imageFactory
-                    .Resize(new Size((int)targetWidth, (int)targetHeight))
-                    .Image;
-                image.Save(imageFile);
-            }
-
-            using (var imageFactory = new ImageProcessor.ImageFactory())
-            {
-                var image = imageFactory
-                    .Load(imageFile)
-                    .GaussianBlur(5)
-                    .Image;
-                image.Save(imageFile);
-            }
-
-            var blurShape = PowerPointCurrentPresentationInfo.CurrentSlide.Shapes.AddPicture(imageFile, Office.MsoTriState.msoFalse,
-                Office.MsoTriState.msoTrue, 0, 0);
-
-            return blurShape;
-        }
-
-        private PowerPoint.Shape GenerateOverlayShape(PowerPoint.Shape shape)
-        {
-            var overlayShape = DuplicateShapeInPlace(shape);
-            
-            overlayShape.Fill.Solid();
-            overlayShape.Fill.ForeColor.RGB = Utils.Graphics.ConvertColorToRgb(Utils.StringUtil.GetColorFromHexValue("#000000"));
-            overlayShape.Fill.Transparency = 80f / 100;
-            overlayShape.Line.ForeColor.RGB = Utils.Graphics.ConvertColorToRgb(Utils.StringUtil.GetColorFromHexValue("#000000"));
-            overlayShape.Line.Transparency = 80f / 100;
-            overlayShape.Line.Weight = 5;
-            overlayShape.Line.Visible = Office.MsoTriState.msoFalse;
-
-            return overlayShape;
-        }
-        
-        private PowerPoint.Shape GenerateShapeFromTextBoundary(PowerPoint.Shape shape)
-        {
-            var rotation = shape.Rotation;
-            if (rotation != 0)
-            {
-                shape.Rotation = 0;
-            }
-
-            var textFrame = shape.TextFrame2;
-            var textRange = textFrame.TextRange.TrimText();
-
-            var left = textRange.BoundLeft - textFrame.MarginLeft;
-            var top = textRange.BoundTop - textFrame.MarginTop;
-            var width = textRange.BoundWidth + textFrame.MarginLeft + textFrame.MarginRight;
-            var height = textRange.BoundHeight + textFrame.MarginTop + textFrame.MarginBottom;
-            
-            var textBoundaryShape = PowerPointCurrentPresentationInfo.CurrentSlide.Shapes.AddShape(Office.MsoAutoShapeType.msoShapeRectangle,
-                left, top, width, height);
-
-            shape.Fill.Visible = Office.MsoTriState.msoFalse;
-            shape.Line.Visible = Office.MsoTriState.msoFalse;
-
-            // for placeholders with text out of box affected due to cut and paste
-            // and text boxes with text out of box
-            if (shape.Height < textBoundaryShape.Height)
-            {
-                shape.Left = textBoundaryShape.Left;
-                shape.Top = textBoundaryShape.Top;
-                shape.Width = textBoundaryShape.Width;
-                shape.Height = textBoundaryShape.Height;
-            }
-
-            if (rotation != 0)
-            {
-                shape.Rotation = rotation;
-                textBoundaryShape.Rotation = rotation;
-            }
-
-            return textBoundaryShape;
-        }
-
-        private PowerPoint.Shape DuplicateShapeInPlace(PowerPoint.Shape shape)
-        {
-            var duplicateShape = shape.Duplicate()[1];
-            duplicateShape.Left = shape.Left;
-            duplicateShape.Top = shape.Top;
-
-            var match = System.Text.RegularExpressions.Regex.Match(duplicateShape.Name, @"\d+$");
-            if (!match.Success || int.Parse(match.Value) != duplicateShape.Id - 1)
-            {
-                duplicateShape.Name += " " + (duplicateShape.Id - 1);
-            }
-
-            return duplicateShape;
         }
         # endregion
